requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
Flask-SocketIO==5.5.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
paho-mqtt==2.1.0
psycopg2-binary==2.9.11
python-engineio==4.12.3
python-socketio==5.15.0
requests==2.32.5
simple-websocket==1.1.0
urllib3==2.6.1
Werkzeug==3.1.4
wsproto==1.3.2
